Linting, fixed bug when package is not installed and added __init__ and __main__ to verbose output
